first major reorganization
