+++ covasim/run.py
@@ -8,7 +8,7 @@ import pylab as pl
 import sciris as sc
 from . import base as cvbase
 from . import model as cvmodel
-import covid_healthsystems as covidhs
+from . import healthsystem as cvhs
 
 
 # Specify all externally visible functions this file defines
@@ -204,7 +204,7 @@ class Scenarios(cvbase.ParsObj):
 
         # Perform health systems analysis
         if healthsystems:
-            self.hsys = covidhs.HealthSystem(self.allres)
+            self.hsys = cvhs.HealthSystem(self.allres)
             self.hsys.analyze()
 
         return
