--- covid_healthsystems/__init__.py
@@ -1,2 +1,0 @@
-from .version import *
-from .healthsystem import *
