--- covid_healthsystems/version.py
@@ -1,4 +1,0 @@
-__all__ = ['__version__', '__versiondate__']
-
-__version__ = '0.1'
-__versiondate__ = '2020-03-16'
